Refactored Neovim config
Deleted old config and started from scratch.
Wrote new config in a more organized manner heavily inspired by
Chris@Machine's config.
Removed a bunch of plugins.
Added new plugins, most importantly vim-which-key and nvim-lspconfig.
Moved to a prerelease version of Neovim to get access to features such
as the built-in LSP client (required for nvim-lspconfig) and tree-sitter
(required for nvim-treesitter).
Created an alias 'cn' for opening Neovim with the environment variables
'GIT_DIR' and 'GIT_WORK_TREE' set to that of the '.dotfiles.git'
repository. This allows git plugins to work with the bare repository,
although not always perfectly. Prior to this, I relied on the plugin to
provide a way to pass arguments to git, which most plugins did not.